Extract area layout and section chrome logic
Refactors DocumentSectionContainer by moving layout state/invariants into a new AreaLayoutState service and gutter/corner calculations into SectionChromeCalculator, then wiring the container to use those components. Adds focused unit tests for both classes to lock in split, visibility, isolation, and chrome behavior, and splits DocumentsPanel tab-menu and utility-docking code into dedicated partial files for clearer structure without changing behavior.
