+++ Source/Workspace/Celbridge.Documents/Views/DocumentsPanel.xaml.cs
@@ -12,13 +12,6 @@ namespace Celbridge.Documents.Views;
 
 using IDocumentsLogger = Logging.ILogger<DocumentsPanel>;
 
-/// <summary>
-/// Where to place a utility when docking it into a document tab. A null Address docks into the active
-/// document's section and appends the tab. A non-null Address targets a specific section and tab order.
-/// Activate selects the docked tab and makes it the active document.
-/// </summary>
-public record DockUtilityPlacement(DocumentAddress? Address, bool Activate);
-
 public sealed partial class DocumentsPanel : UserControl, IDocumentsPanel
 {
     private readonly IServiceProvider _serviceProvider;
@@ -803,80 +796,6 @@ public sealed partial class DocumentsPanel : UserControl, IDocumentsPanel
         return Result.Ok();
     }
 
-    /// <summary>
-    /// Docks a utility into a document tab: creates a tab hosting the utility's borrowed controller (reusing its
-    /// live WebView) and stamps the utility tab metadata. The controller's WebView is reparented into the tab
-    /// once it is in the visual tree.
-    /// </summary>
-    public Result DockUtility(CustomUtilityView panelView, DockUtilityPlacement placement)
-    {
-        var resource = panelView.FileResource;
-        var editorId = panelView.UtilityId;
-
-        var resolveResult = ViewModel.ResolveResourcePath(resource);
-        if (resolveResult.IsFailure)
-        {
-            return Result.Fail($"Failed to resolve path for utility resource: '{resource}'")
-                .WithErrors(resolveResult);
-        }
-        var filePath = resolveResult.Value;
-
-        var address = placement.Address;
-        var section = address is not null ? EnsureSectionMounted(address.Section) : SectionContainer.ActiveSection;
-        var sectionView = SectionContainer.GetSection(section);
-
-        var documentTab = new DocumentTab();
-        documentTab.ViewModel.FileResource = resource;
-        documentTab.ViewModel.FilePath = filePath;
-        documentTab.ViewModel.EditorId = editorId;
-        ApplyUtilityTabMetadata(documentTab, editorId);
-
-        var dockedView = new DockedUtilityDocumentView(_serviceProvider, _messengerService, panelView.Controller);
-        dockedView.EditorId = editorId;
-        dockedView.Bind(resource, filePath);
-
-        if (address is not null)
-        {
-            sectionView.InsertTab(documentTab, address.TabOrder);
-        }
-        else
-        {
-            sectionView.AddTab(documentTab);
-        }
-
-        // No open announcement: a utility is presented by docking, never opened as a document, and the
-        // documents service refuses to open one. The view model suppresses the matching close.
-        documentTab.ViewModel.DocumentView = dockedView;
-        documentTab.Content = dockedView;
-
-        if (placement.Activate)
-        {
-            sectionView.SelectTab(documentTab);
-        }
-
-        // Reparent the borrowed WebView into the tab now that the tab is in the visual tree.
-        dockedView.Dock();
-
-        if (placement.Activate)
-        {
-            SectionContainer.ActivateDocument(resource, section);
-        }
-
-        return Result.Ok();
-    }
-
-    /// <summary>
-    /// Activates the open document tab for a docked utility (used when its rail button is clicked).
-    /// </summary>
-    public void ActivateUtilityTab(ResourceKey resource)
-    {
-        var location = SectionContainer.FindDocumentTab(resource);
-        if (location is not null)
-        {
-            SectionContainer.ActivateDocument(resource, location.SectionView.Section);
-        }
-    }
-
     private void OnFlashDocumentRequested(object recipient, FlashDocumentMessage message)
     {
         // Flashing is a transient view effect, so it is deferred until the tab that prompted it (a freshly
@@ -895,30 +814,6 @@ public sealed partial class DocumentsPanel : UserControl, IDocumentsPanel
     {
         var location = SectionContainer.FindDocumentTab(fileResource);
         location?.Tab.FlashAttention();
-    }
-
-    /// <summary>
-    /// Removes a docked utility's document tab. The caller reparents the controller's WebView back to the
-    /// Utility Panel first, so the tab (and its now-empty docked view) is dropped without any teardown.
-    /// </summary>
-    public void RemoveUtilityTab(ResourceKey resource)
-    {
-        var location = SectionContainer.FindDocumentTab(resource);
-        if (location is null)
-        {
-            return;
-        }
-
-        var sectionView = location.SectionView;
-        var documentTab = location.Tab;
-
-        int tabIndex = sectionView.GetTabIndex(documentTab);
-        SectionContainer.HandleDocumentClosing(resource, sectionView.Section, tabIndex);
-
-        _ = documentTab.ViewModel.DocumentView?.PrepareToClose();
-        RemoveTabFromSection(sectionView, documentTab);
-
-        UpdateAllTabDisplayNames();
     }
 
     // Removes a tab, folding its area back when that leaves one of a split area's sections empty.
@@ -1149,20 +1044,6 @@ public sealed partial class DocumentsPanel : UserControl, IDocumentsPanel
         }
     }
 
-    private void ApplyUtilityTabMetadata(DocumentTab documentTab, EditorId editorId)
-    {
-        var utilityInfo = ViewModel.ResolveUtilityTabInfo(editorId);
-        if (utilityInfo is null)
-        {
-            return;
-        }
-
-        documentTab.ViewModel.IsUtility = true;
-        documentTab.ViewModel.UtilityIconName = utilityInfo.IconName;
-        documentTab.ViewModel.DocumentName = utilityInfo.Title;
-        documentTab.ViewModel.UtilityTooltip = utilityInfo.Tooltip;
-    }
-
     private void UpdateAllTabDisplayNames()
     {
         // Collect all tabs from all sections. Utility tabs keep their manifest title, so they are
@@ -1222,332 +1103,9 @@ public sealed partial class DocumentsPanel : UserControl, IDocumentsPanel
         SectionContainer.Shutdown();
     }
 
-    private void OnDocumentTabContextMenuAction(DocumentTab tab, DocumentTabMenuAction action)
-    {
-        switch (action)
-        {
-            case DocumentTabMenuAction.Close:
-                CloseTab(tab);
-                break;
-            case DocumentTabMenuAction.CloseOthers:
-                CloseOtherTabs(tab);
-                break;
-            case DocumentTabMenuAction.CloseOthersRight:
-                CloseOtherTabsRight(tab);
-                break;
-            case DocumentTabMenuAction.CloseOthersLeft:
-                CloseOtherTabsLeft(tab);
-                break;
-            case DocumentTabMenuAction.CloseAll:
-                CloseAllTabs(tab);
-                break;
-            case DocumentTabMenuAction.MoveToPrimarySection:
-                MoveTabWithinArea(tab, toSecondarySection: false);
-                break;
-            case DocumentTabMenuAction.MoveToSecondarySection:
-                MoveTabWithinArea(tab, toSecondarySection: true);
-                break;
-            case DocumentTabMenuAction.UnsplitArea:
-                UnsplitArea(tab);
-                break;
-            case DocumentTabMenuAction.CopyResourceKey:
-                CopyResourceKeyForTab(tab);
-                break;
-            case DocumentTabMenuAction.CopyFilePath:
-                CopyFilePathForTab(tab);
-                break;
-            case DocumentTabMenuAction.SelectFile:
-                SelectFileForTab(tab);
-                break;
-            case DocumentTabMenuAction.OpenFileExplorer:
-                OpenFileExplorerForTab(tab);
-                break;
-            case DocumentTabMenuAction.OpenApplication:
-                OpenApplicationForTab(tab);
-                break;
-            case DocumentTabMenuAction.RestoreChrome:
-                RestoreChromeForTab(tab);
-                break;
-            case DocumentTabMenuAction.Reopen:
-                _ = ReopenTab(tab);
-                break;
-            case DocumentTabMenuAction.ReopenWith:
-                _ = ReopenTabWithDialog(tab);
-                break;
-        }
-    }
-
-    private void CloseTab(DocumentTab tab)
-    {
-        var fileResource = tab.ViewModel.FileResource;
-        ViewModel.OnCloseDocumentRequested(fileResource);
-    }
-
-    // Moves a tab between the two sections of its own area, splitting the area first when it is not split
-    // yet. Moving between areas is a drag.
-    private void MoveTabWithinArea(DocumentTab tab, bool toSecondarySection)
-    {
-        var area = tab.Section.GetArea();
-
-        if (!SectionContainer.IsAreaSplit(area))
-        {
-            // Only the split direction is on offer while unsplit, and only while the area has a document
-            // to leave behind.
-            if (!toSecondarySection ||
-                !SectionContainer.CanStartAreaSplit(area))
-            {
-                return;
-            }
-
-            SectionContainer.SetAreaSplit(area, true);
-        }
-
-        var targetSection = toSecondarySection
-            ? area.GetSecondarySection()
-            : area.GetPrimarySection();
-
-        if (SectionContainer.MoveTabToSection(tab, targetSection))
-        {
-            UpdateAllTabDisplayNames();
-            NotifyLayoutChanged();
-        }
-    }
-
-    // Folds a split area back, merging both sections into the primary one.
-    private void UnsplitArea(DocumentTab tab)
-    {
-        var area = tab.Section.GetArea();
-        if (!SectionContainer.IsAreaSplit(area))
-        {
-            return;
-        }
-
-        SectionContainer.SetAreaSplit(area, false);
-
-        UpdateAllTabDisplayNames();
-        NotifyLayoutChanged();
-    }
-
     private void NotifyLayoutChanged()
     {
         ViewModel.OnDocumentLayoutChanged();
     }
 
-    private void CloseOtherTabs(DocumentTab keepTab)
-    {
-        // Find which section contains the tab to keep
-        var location = SectionContainer.FindDocumentTab(keepTab.ViewModel.FileResource);
-        if (location is null)
-        {
-            return;
-        }
-
-        var sectionView = location.SectionView;
-
-        var tabsToClose = new List<ResourceKey>();
-
-        // Only close other tabs within the same section.
-        foreach (var documentTab in sectionView.GetAllTabs())
-        {
-            if (documentTab != keepTab)
-            {
-                tabsToClose.Add(documentTab.ViewModel.FileResource);
-            }
-        }
-
-        foreach (var fileResource in tabsToClose)
-        {
-            ViewModel.OnCloseDocumentRequested(fileResource);
-        }
-    }
-
-    private void CloseOtherTabsRight(DocumentTab referenceTab)
-    {
-        // Find which section contains the reference tab
-        var location = SectionContainer.FindDocumentTab(referenceTab.ViewModel.FileResource);
-        if (location is null)
-        {
-            return;
-        }
-
-        var sectionView = location.SectionView;
-
-        var tabsToClose = new List<ResourceKey>();
-        bool foundReference = false;
-
-        // Close tabs to the right within the same section.
-        foreach (var documentTab in sectionView.GetAllTabs())
-        {
-            if (foundReference)
-            {
-                tabsToClose.Add(documentTab.ViewModel.FileResource);
-            }
-            if (documentTab == referenceTab)
-            {
-                foundReference = true;
-            }
-        }
-
-        foreach (var fileResource in tabsToClose)
-        {
-            ViewModel.OnCloseDocumentRequested(fileResource);
-        }
-    }
-
-    private void CloseOtherTabsLeft(DocumentTab referenceTab)
-    {
-        // Find which section contains the reference tab
-        var location = SectionContainer.FindDocumentTab(referenceTab.ViewModel.FileResource);
-        if (location is null)
-        {
-            return;
-        }
-
-        var sectionView = location.SectionView;
-
-        var tabsToClose = new List<ResourceKey>();
-
-        // Close tabs to the left within the same section.
-        foreach (var documentTab in sectionView.GetAllTabs())
-        {
-            if (documentTab == referenceTab)
-            {
-                break;
-            }
-            tabsToClose.Add(documentTab.ViewModel.FileResource);
-        }
-
-        foreach (var fileResource in tabsToClose)
-        {
-            ViewModel.OnCloseDocumentRequested(fileResource);
-        }
-    }
-
-    private void CloseAllTabs(DocumentTab referenceTab)
-    {
-        // Find which section contains the reference tab
-        var location = SectionContainer.FindDocumentTab(referenceTab.ViewModel.FileResource);
-        if (location is null)
-        {
-            return;
-        }
-
-        var sectionView = location.SectionView;
-
-        var tabsToClose = new List<ResourceKey>();
-
-        // Only close tabs within the same section.
-        foreach (var documentTab in sectionView.GetAllTabs())
-        {
-            tabsToClose.Add(documentTab.ViewModel.FileResource);
-        }
-
-        foreach (var fileResource in tabsToClose)
-        {
-            ViewModel.OnCloseDocumentRequested(fileResource);
-        }
-    }
-
-    private void SelectFileForTab(DocumentTab tab)
-    {
-        ViewModel.SelectFileForTab(tab.ViewModel.FileResource);
-    }
-
-    private void CopyResourceKeyForTab(DocumentTab tab)
-    {
-        ViewModel.CopyResourceKeyForTab(tab.ViewModel.FileResource);
-    }
-
-    private void CopyFilePathForTab(DocumentTab tab)
-    {
-        ViewModel.CopyFilePathForTab(tab.ViewModel.FilePath);
-    }
-
-    private void OpenFileExplorerForTab(DocumentTab tab)
-    {
-        ViewModel.OpenFileExplorerForTab(tab.ViewModel.FileResource);
-    }
-
-    private void OpenApplicationForTab(DocumentTab tab)
-    {
-        ViewModel.OpenApplicationForTab(tab.ViewModel.FileResource);
-    }
-
-    private void RestoreChromeForTab(DocumentTab tab)
-    {
-        if (tab.Content is IDocumentChromeOwner chromeOwner)
-        {
-            chromeOwner.RestoreChrome();
-        }
-    }
-
-    private Task ReopenTab(DocumentTab tab)
-    {
-        // Reopen using the current editor (no dialog)
-        return ReopenTabWithEditor(tab, tab.ViewModel.EditorId);
-    }
-
-    private async Task ReopenTabWithDialog(DocumentTab tab)
-    {
-        var fileResource = tab.ViewModel.FileResource;
-
-        var selectedEditorId = tab.ViewModel.EditorId;
-
-        var pickList = ViewModel.GetEditorPickList(fileResource, tab.ViewModel.EditorId);
-        if (pickList is not null)
-        {
-            // Multiple editors available, show choice dialog.
-            var title = _stringLocalizer.GetString("OpenWithDialog_Title");
-            var message = _stringLocalizer.GetString("OpenWithDialog_Message");
-
-            var choiceResult = await _dialogService.ShowChoiceDialogAsync(
-                title, message, pickList.Labels, pickList.SelectedIndex, checkbox: null);
-            if (choiceResult.IsFailure)
-            {
-                return;
-            }
-
-            selectedEditorId = pickList.EditorIds[choiceResult.Value.SelectedIndex];
-
-            await ViewModel.SetPreferredEditorAsync(fileResource, selectedEditorId);
-        }
-
-        await ReopenTabWithEditor(tab, selectedEditorId);
-    }
-
-    private async Task ReopenTabWithEditor(DocumentTab tab, EditorId editorId)
-    {
-        var fileResource = tab.ViewModel.FileResource;
-
-        // Capture state before closing so we can restore it after reopening
-        var section = tab.Section;
-        var currentLocation = SectionContainer.FindDocumentTab(fileResource);
-        var tabIndex = currentLocation?.SectionView.GetTabIndex(tab) ?? 0;
-
-        string? editorState = null;
-        if (tab.ViewModel.DocumentView is not null)
-        {
-            editorState = await tab.ViewModel.DocumentView.TrySaveEditorStateAsync();
-        }
-
-        // Close then reopen via the command service, which processes them sequentially
-        var closeResult = await _commandService.ExecuteAsync<ICloseDocumentCommand>(command =>
-        {
-            command.FileResource = fileResource;
-        });
-
-        if (closeResult.IsFailure)
-        {
-            return;
-        }
-
-        _commandService.Execute<IOpenDocumentCommand>(command =>
-        {
-            command.FileResource = fileResource;
-            command.EditorId = editorId;
-            command.EditorStateJson = editorState;
-            command.TargetSection = section;
-            command.TargetTabIndex = tabIndex;
-        });
-    }
 }
